Move beta things to cmd/crank/beta
