+++ cmd/crank/beta/render/testdata/xr.yaml
