+++ cmd/crank/xpkg/install.go
@@ -36,9 +36,7 @@ import (
 	"github.com/crossplane/crossplane-runtime/pkg/logging"
 
 	v1 "github.com/crossplane/crossplane/apis/pkg/v1"
-	"github.com/crossplane/crossplane/apis/pkg/v1beta1"
 	clientpkgv1 "github.com/crossplane/crossplane/internal/client/clientset/versioned/typed/pkg/v1"
-	clientpkgv1beta1 "github.com/crossplane/crossplane/internal/client/clientset/versioned/typed/pkg/v1beta1"
 	"github.com/crossplane/crossplane/internal/version"
 	"github.com/crossplane/crossplane/internal/xpkg"
 
@@ -63,10 +61,6 @@ const (
 	msgProviderReady    = "Provider is ready"
 	msgProviderNotReady = "Provider is not ready"
 	msgProviderWaiting  = "Waiting for the Provider to be ready"
-
-	msgFunctionReady    = "Function is ready"
-	msgFunctionNotReady = "Function is not ready"
-	msgFunctionWaiting  = "Waiting for the Function to be ready"
 )
 
 // TODO(negz): These install<T>Cmd implementations are all identical. Can they
@@ -75,7 +69,6 @@ const (
 type installCmd struct {
 	Configuration installConfigCmd   `cmd:"" help:"Install a Configuration package."`
 	Provider      installProviderCmd `cmd:"" help:"Install a Provider package."`
-	Function      installFunctionCmd `cmd:"" help:"Install a Function package."`
 }
 
 // installConfigCmd installs a Configuration.
@@ -265,104 +258,6 @@ func (c *installProviderCmd) Run(k *kong.Context, logger logging.Logger) error {
 		}
 	}
 	_, err = fmt.Fprintf(k.Stdout, "%s/%s created\n", strings.ToLower(v1.ProviderGroupKind), res.GetName())
-	return err
-}
-
-// installFunctionCmd installs a Function.
-type installFunctionCmd struct {
-	Package string `arg:"" help:"Image containing Function package."`
-
-	Name                 string        `arg:"" optional:"" help:"Name of Function."`
-	Wait                 time.Duration `short:"w" help:"Wait for installation of package"`
-	RevisionHistoryLimit int64         `short:"r" help:"Revision history limit."`
-	ManualActivation     bool          `short:"m" help:"Enable manual revision activation policy."`
-	Config               string        `help:"Specify a DeploymentRuntimeConfig for this Function."`
-	PackagePullSecrets   []string      `help:"List of secrets used to pull package."`
-}
-
-// Run runs the Function install cmd.
-func (c *installFunctionCmd) Run(k *kong.Context, logger logging.Logger) error { //nolint:gocyclo // TODO(negz): Can anything be broken out here?
-	rap := v1.AutomaticActivation
-	if c.ManualActivation {
-		rap = v1.ManualActivation
-	}
-	pkgName := c.Name
-	if pkgName == "" {
-		ref, err := name.ParseReference(c.Package)
-		if err != nil {
-			logger.Debug(errPkgIdentifier, "error", err)
-			return errors.Wrap(err, errPkgIdentifier)
-		}
-		pkgName = xpkg.ToDNSLabel(ref.Context().RepositoryStr())
-	}
-	logger = logger.WithValues("providerName", pkgName)
-	packagePullSecrets := make([]corev1.LocalObjectReference, len(c.PackagePullSecrets))
-	for i, s := range c.PackagePullSecrets {
-		packagePullSecrets[i] = corev1.LocalObjectReference{
-			Name: s,
-		}
-	}
-	cr := &v1beta1.Function{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: pkgName,
-		},
-		Spec: v1beta1.FunctionSpec{
-			PackageSpec: v1.PackageSpec{
-				Package:                  c.Package,
-				RevisionActivationPolicy: &rap,
-				RevisionHistoryLimit:     &c.RevisionHistoryLimit,
-				PackagePullSecrets:       packagePullSecrets,
-			},
-		},
-	}
-	if c.Config != "" {
-		cr.Spec.ControllerConfigReference = &v1.ControllerConfigReference{
-			Name: c.Config,
-		}
-	}
-	kubeConfig, err := ctrl.GetConfig()
-	if err != nil {
-		logger.Debug(errKubeConfig, "error", err)
-		return errors.Wrap(err, errKubeConfig)
-	}
-	logger.Debug("Found kubeconfig")
-	kube, err := clientpkgv1beta1.NewForConfig(kubeConfig)
-	if err != nil {
-		logger.Debug(errKubeClient, "error", err)
-		return errors.Wrap(err, errKubeClient)
-	}
-	logger.Debug("Created kubernetes client")
-	res, err := kube.Functions().Create(context.Background(), cr, metav1.CreateOptions{})
-	if err != nil {
-		logger.Debug("Failed to create function", "error", warnIfNotFound(err))
-		return errors.Wrap(warnIfNotFound(err), "cannot create function")
-	}
-	if c.Wait != 0 {
-		logger.Debug(msgFunctionWaiting)
-		watchList := cache.NewListWatchFromClient(kube.RESTClient(), "functions", corev1.NamespaceAll, fields.Everything())
-		waitSeconds := int64(c.Wait.Seconds())
-		watcher, err := watchList.Watch(metav1.ListOptions{Watch: true, TimeoutSeconds: &waitSeconds})
-		defer watcher.Stop()
-		if err != nil {
-			logger.Debug(fmt.Sprintf(errFmtWatchPkg, "Function"), "error", err)
-			return err
-		}
-		for {
-			event, ok := <-watcher.ResultChan()
-			if !ok {
-				logger.Debug(fmt.Sprintf(errFmtPkgNotReadyTimeout, "Function"))
-				return errors.Errorf(errFmtPkgNotReadyTimeout, "Function")
-			}
-			obj := (event.Object).(*v1beta1.Function)
-			cond := obj.GetCondition(v1.TypeHealthy)
-			if obj.ObjectMeta.Name == pkgName && cond.Status == corev1.ConditionTrue {
-				logger.Debug(msgFunctionReady, "pkgName", obj.ObjectMeta.Name)
-				break
-			}
-			logger.Debug(msgFunctionNotReady)
-		}
-	}
-	_, err = fmt.Fprintf(k.Stdout, "%s/%s created\n", strings.ToLower(v1beta1.FunctionGroupKind), res.GetName())
 	return err
 }
