+++ cmd/crank/xpkg/update.go
@@ -32,9 +32,7 @@ import (
 	"github.com/crossplane/crossplane-runtime/pkg/logging"
 
 	v1 "github.com/crossplane/crossplane/apis/pkg/v1"
-	"github.com/crossplane/crossplane/apis/pkg/v1beta1"
 	clientpkgv1 "github.com/crossplane/crossplane/internal/client/clientset/versioned/typed/pkg/v1"
-	clientpkgv1beta1 "github.com/crossplane/crossplane/internal/client/clientset/versioned/typed/pkg/v1beta1"
 
 	// Load all the auth plugins for the cloud providers.
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
@@ -44,7 +42,6 @@ import (
 type updateCmd struct {
 	Configuration updateConfigCmd   `cmd:"" help:"Update a Configuration package."`
 	Provider      updateProviderCmd `cmd:"" help:"Update a Provider package."`
-	Function      updateFunctionCmd `cmd:"" help:"Update a Function package."`
 }
 
 // updateConfigCmd updates a Configuration.
@@ -159,62 +156,5 @@ func (c *updateProviderCmd) Run(k *kong.Context, logger logging.Logger) error {
 		return errors.Wrap(err, "cannot update provider")
 	}
 	_, err = fmt.Fprintf(k.Stdout, "%s/%s updated\n", strings.ToLower(v1.ProviderGroupKind), res.GetName())
-	return err
-}
-
-// updateFunctionCmd update a Function.
-type updateFunctionCmd struct {
-	Name string `arg:"" help:"Name of Function."`
-	Tag  string `arg:"" help:"Updated tag for Function package."`
-}
-
-// Run runs the Function update cmd.
-func (c *updateFunctionCmd) Run(k *kong.Context, logger logging.Logger) error {
-	kubeConfig, err := ctrl.GetConfig()
-	if err != nil {
-		logger.Debug(errKubeConfig, "error", err)
-		return errors.Wrap(err, errKubeConfig)
-	}
-	logger.Debug("Found kubeconfig")
-	kube, err := clientpkgv1beta1.NewForConfig(kubeConfig)
-	if err != nil {
-		logger.Debug(errKubeClient, "error", err)
-		return errors.Wrap(err, errKubeClient)
-	}
-	logger.Debug("Created kubernetes client")
-	preProv, err := kube.Functions().Get(context.Background(), c.Name, metav1.GetOptions{})
-	if err != nil {
-		err = warnIfNotFound(err)
-		logger.Debug("Failed to update function", "error", err)
-		return errors.Wrap(err, "cannot update function")
-	}
-	logger.Debug("Found previous function object")
-	pkg := preProv.Spec.Package
-	pkgReference, err := name.ParseReference(pkg, name.WithDefaultRegistry(""))
-	if err != nil {
-		err = warnIfNotFound(err)
-		logger.Debug("Failed to update function", "error", err)
-		return errors.Wrap(err, "cannot update function")
-	}
-	newPkg := ""
-	if strings.HasPrefix(c.Tag, "sha256") {
-		newPkg = pkgReference.Context().Digest(c.Tag).Name()
-	} else {
-		newPkg = pkgReference.Context().Tag(c.Tag).Name()
-	}
-	preProv.Spec.Package = newPkg
-	req, err := json.Marshal(preProv)
-	if err != nil {
-		err = warnIfNotFound(err)
-		logger.Debug("Failed to update function", "error", err)
-		return errors.Wrap(err, "cannot update function")
-	}
-	res, err := kube.Functions().Patch(context.Background(), c.Name, types.MergePatchType, req, metav1.PatchOptions{})
-	if err != nil {
-		err = warnIfNotFound(err)
-		logger.Debug("Failed to update function", "error", err)
-		return errors.Wrap(err, "cannot update function")
-	}
-	_, err = fmt.Fprintf(k.Stdout, "%s/%s updated\n", strings.ToLower(v1beta1.FunctionGroupKind), res.GetName())
 	return err
 }
