+++ cmd/crank/xpkg/xpkg.go
@@ -14,18 +14,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package xpkg contains the Crossplane packaging commands.
+// Package xpkg contains Crossplane packaging commands.
 package xpkg
 
 // Cmd contains commands for interacting with xpkgs.
 // TODO(lsviben) add the rest of the commands from up (batch, xpextract).
 type Cmd struct {
 	Build   buildCmd   `cmd:"" help:"Build a package, by default from the current directory."`
-	Push    pushCmd    `cmd:"" help:"Push Crossplane packages."`
+	Push    pushCmd    `cmd:"" help:"Push a package."`
 	Login   loginCmd   `cmd:"" help:"Login to the default package registry (xpkg.upbound.io)."`
 	Logout  logoutCmd  `cmd:"" help:"Logout of the default package registry (xpkg.upbound.io)."`
-	Install installCmd `cmd:"" help:"Install Crossplane packages."`
-	Update  updateCmd  `cmd:"" help:"Update Crossplane packages."`
+	Install installCmd `cmd:"" help:"Install a package."`
+	Update  updateCmd  `cmd:"" help:"Update an installed package."`
 }
 
 // Help prints out the help for the xpkg command.
